requirements.txt
beautifulsoup4
lxml
requests
pandas
httpx
python-dotenv